+++ migrations/0029_secret_fantastic_four.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "playlists" ADD COLUMN "show_next_button" boolean DEFAULT false NOT NULL;